+++ tamr_unify_client/client.py
@@ -4,7 +4,7 @@ import requests
 from requests import Response
 
 from tamr_unify_client.dataset.collection import DatasetCollection
-from tamr_unify_client.models.project.collection import ProjectCollection
+from tamr_unify_client.project.collection import ProjectCollection
 
 # monkey-patch Response.successful
 
@@ -138,7 +138,7 @@ class Client:
         """Collection of all projects on this Unify instance.
 
         :return: Collection of all projects.
-        :rtype: :class:`~tamr_unify_client.models.ProjectCollection`
+        :rtype: :class:`~tamr_unify_client.project.collection.ProjectCollection`
         """
         return self._projects
