+++ tests/unit/test_dataset_usage.py
@@ -7,7 +7,7 @@ from tamr_unify_client.auth import UsernamePasswordAuth
 from tamr_unify_client.dataset.resource import Dataset
 from tamr_unify_client.dataset.usage import DatasetUsage
 from tamr_unify_client.dataset.use import DatasetUse
-from tamr_unify_client.models.project.step import ProjectStep
+from tamr_unify_client.project.step import ProjectStep
 
 
 class TestUsage(TestCase):
