+++ tests/unit/test_published_clusters.py
@@ -4,10 +4,10 @@ import responses
 
 from tamr_unify_client import Client
 from tamr_unify_client.auth import UsernamePasswordAuth
-from tamr_unify_client.models.project.cluster_configuration import (
+from tamr_unify_client.project.cluster_configuration import (
     PublishedClustersConfiguration,
 )
-from tamr_unify_client.models.project.resource import Project
+from tamr_unify_client.project.resource import Project
 
 
 class PublishedClusterTest(TestCase):
